rename de package assetFile for asset
