+++ src/test/java/com/jala/test/SearchFileTest.java
@@ -13,7 +13,7 @@
 package com.jala.test;
 
 import com.jala.model.search.SearchFile;
-import com.jala.model.search.assetFile.Asset;
+import com.jala.model.search.asset.Asset;
 import com.jala.model.criteria.CriteriaSearch;
 import com.jala.model.search.TernaryBooleanEnum;
 import org.junit.After;
